Updated IK. Hands will lerp to target when interaction begins, and will lerp back when interaction finishes.

+++ Unity/Assets/Scripts/Player/CPlayerGroundMotor.cs
@@ -640,7 +640,7 @@ public class CPlayerGroundMotor : CNetworkMonoBehaviour
                 // Update state
                 if (_bDown)
                 {
-                    m_uiInputStates |= (uint)eTargetState;
+                    m_uiInputStates |= (ushort)eTargetState;
                 }
                 else
                 {
